Improve ACPI implementation
Corrects the RSDP structure as it was written for ACPI 2.0+ and the rest of the code only realistically supports ACPI 1.0 which is the only supported version of ACPI in QEMU SeaBIOS. Reading any of the additional fields would result in undefined behavior and would yield garbage.
Adds some sanity checks in the form of asserts (e.g. makes sure the RSDP is actually valid). Uses ptr::read_unaligned instead of directly trying to dereference the pointer to RSDT entries because as it turns out, those tables can be stored at misaligned addresses.